eval raw parameters that eval to something raw
if one function takes :rawLines & passes it to another function that
takes :rawLines, we need to eval it before passing it to the second
function.  if we don't, we end up wrapping the raw values in an
extra unnecessary layer.

this allows elsif to take :rawLines & pass it to &?, which also takes
:rawLines