fixup tests so they can find the l3 directory
having 'unittest' fail when it couldn't find files pointed out that
several of the tests weren't running because the current dir wasn't set
to anyplace where the l3 directory could be found.
instead of copying the l3 files (which made debugging a pain & wasn't
sufficient anyway), switch the working directory to the main app
directory where the 'l3' directory is.